# Conflicts:
#	CONTRIBUTORS
#	README.md
#	pom.xml
#	xchange-abucoins/pom.xml
#	xchange-acx/pom.xml
#	xchange-anx/pom.xml
#	xchange-bibox/pom.xml
#	xchange-binance/pom.xml
#	xchange-binance/src/main/java/org/knowm/xchange/binance/service/BinanceTradeService.java
#	xchange-bitbay/pom.xml
#	xchange-bitcoinaverage/pom.xml
#	xchange-bitcoincharts/pom.xml
#	xchange-bitcoincore/pom.xml
#	xchange-bitcoinde/pom.xml
#	xchange-bitcoinium/pom.xml
#	xchange-bitcointoyou/pom.xml
#	xchange-bitfinex/pom.xml
#	xchange-bitfinex/src/main/java/org/knowm/xchange/bitfinex/v2/BitfinexExchange.java
#	xchange-bitflyer/pom.xml
#	xchange-bitmarket/pom.xml
#	xchange-bitmex/pom.xml
#	xchange-bitmex/src/main/java/org/knowm/xchange/bitmex/Bitmex.java
#	xchange-bitmex/src/main/java/org/knowm/xchange/bitmex/service/BitmexAccountServiceRaw.java
#	xchange-bitmex/src/main/java/org/knowm/xchange/bitmex/service/BitmexBaseService.java
#	xchange-bitmex/src/main/java/org/knowm/xchange/bitmex/service/BitmexMarketDataServiceRaw.java
#	xchange-bitmex/src/main/java/org/knowm/xchange/bitmex/service/BitmexTradeService.java
#	xchange-bitmex/src/main/java/org/knowm/xchange/bitmex/service/BitmexTradeServiceRaw.java
#	xchange-bitmex/src/test/java/org/knowm/xchange/bitmex/service/BitmexBulkOrderTest.java
#	xchange-bitso/pom.xml
#	xchange-bitstamp/pom.xml
#	xchange-bitstamp/src/main/java/org/knowm/xchange/bitstamp/BitstampAuthenticatedV2.java
#	xchange-bitstamp/src/main/java/org/knowm/xchange/bitstamp/dto/account/BitstampRippleDepositAddress.java
#	xchange-bitstamp/src/main/java/org/knowm/xchange/bitstamp/dto/account/WithdrawalRequest.java
#	xchange-bitstamp/src/main/java/org/knowm/xchange/bitstamp/service/BitstampAccountServiceRaw.java
#	xchange-bittrex/pom.xml
#	xchange-bity/pom.xml
#	xchange-bitz/pom.xml
#	xchange-bleutrade/pom.xml
#	xchange-blockchain/pom.xml
#	xchange-btcc/pom.xml
#	xchange-btcmarkets/pom.xml
#	xchange-btctrade/pom.xml
#	xchange-btcturk/pom.xml
#	xchange-bx/pom.xml
#	xchange-campbx/pom.xml
#	xchange-ccex/pom.xml
#	xchange-cexio/pom.xml
#	xchange-cobinhood/pom.xml
#	xchange-coinbase/pom.xml
#	xchange-coinbene/pom.xml
#	xchange-coindirect/pom.xml
#	xchange-coinegg/pom.xml
#	xchange-coinfloor/pom.xml
#	xchange-coingi/pom.xml
#	xchange-coinmarketcap/pom.xml
#	xchange-coinmate/pom.xml
#	xchange-coinone/pom.xml
#	xchange-coinsuper/pom.xml
#	xchange-core/pom.xml
#	xchange-cryptofacilities/pom.xml
#	xchange-cryptonit/pom.xml
#	xchange-cryptopia/pom.xml
#	xchange-dsx/pom.xml
#	xchange-examples/pom.xml
#	xchange-exmo/pom.xml
#	xchange-exx/pom.xml
#	xchange-fcoin/pom.xml
#	xchange-gatecoin/pom.xml
#	xchange-gateio/pom.xml
#	xchange-gdax/pom.xml
#	xchange-gemini/pom.xml
#	xchange-hitbtc/pom.xml
#	xchange-huobi/pom.xml
#	xchange-huobi/src/main/java/org/knowm/xchange/huobi/HuobiAdapters.java
#	xchange-idex/pom.xml
#	xchange-independentreserve/pom.xml
#	xchange-itbit/pom.xml
#	xchange-koineks/pom.xml
#	xchange-koinim/pom.xml
#	xchange-kraken/pom.xml
#	xchange-kucoin/pom.xml
#	xchange-kuna/pom.xml
#	xchange-lakebtc/pom.xml
#	xchange-liqui/pom.xml
#	xchange-livecoin/pom.xml
#	xchange-luno/pom.xml
#	xchange-mercadobitcoin/pom.xml
#	xchange-okcoin/pom.xml
#	xchange-openexchangerates/pom.xml
#	xchange-paribu/pom.xml
#	xchange-paymium/pom.xml
#	xchange-poloniex/pom.xml
#	xchange-quadrigacx/pom.xml
#	xchange-quoine/pom.xml
#	xchange-ripple/pom.xml
#	xchange-therock/pom.xml
#	xchange-truefx/pom.xml
#	xchange-upbit/pom.xml
#	xchange-upbit/src/main/java/org/knowm/xchange/upbit/UpbitAdapters.java
#	xchange-upbit/src/main/java/org/knowm/xchange/upbit/UpbitAuthenticated.java
#	xchange-upbit/src/main/java/org/knowm/xchange/upbit/UpbitExchange.java
#	xchange-upbit/src/main/java/org/knowm/xchange/upbit/service/UpbitBaseService.java
#	xchange-vaultoro/pom.xml
#	xchange-vircurex/pom.xml
#	xchange-wex/pom.xml
#	xchange-yobit/pom.xml
#	xchange-zaif/pom.xml

+++ xchange-bitstamp/src/main/java/org/knowm/xchange/bitstamp/BitstampAuthenticated.java
@@ -72,7 +72,7 @@ public interface BitstampAuthenticated {
       @FormParam("key") String apiKey,
       @FormParam("signature") ParamsDigest signer,
       @FormParam("nonce") SynchronizedValueFactory<Long> nonce,
-      @FormParam("id") int orderId)
+      @FormParam("id") long orderId)
       throws BitstampException, IOException;
 
   /** @return true if order has been canceled. */
